Rename InterfaceAdapter.change_password to InterfaceAdapter.logged_in
