+++ src/main/java/app/AppBuilder.java
@@ -1,6 +1,5 @@
 package app;
 
-import java.Constants;
 import java.awt.*;
 
 import javax.swing.*;
@@ -9,7 +8,7 @@ import data_access.InMemoryUserDataAccessObject;
 import entity.CommonUserFactory;
 import entity.UserFactory;
 import interface_adapter.ViewManagerModel;
-import interface_adapter.change_password.LoggedInViewModel;
+import interface_adapter.logged_in.LoggedInViewModel;
 import interface_adapter.login.LoginController;
 import interface_adapter.login.LoginPresenter;
 import interface_adapter.login.LoginViewModel;
