+++ src/main/java/interface_adapter/logged_in/LoggedInState.java
@@ -1,4 +1,4 @@
-package interface_adapter.change_password;
+package interface_adapter.logged_in;
 
 /**
  * The State information representing the logged-in user.
